% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{macrox_server}
\alias{macrox_server}
\title{Shiny module server for PDF table extraction}
\usage{
macrox_server(id)
}
\arguments{
\item{id}{Module namespace ID (must match \code{\link[=macrox_ui]{macrox_ui()}}).}
}
\value{
A list with two reactives: \code{tables} (named list of data frames) and
\code{steps} (list of step definitions).
}
\description{
Shiny module server for PDF table extraction
}
