% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{pdfmacro_server}
\alias{pdfmacro_server}
\title{Shiny module server for PDF table extraction}
\usage{
pdfmacro_server(id)
}
\arguments{
\item{id}{Module namespace ID (must match \code{\link[=pdfmacro_ui]{pdfmacro_ui()}}).}
}
\value{
A list with two reactives: \code{tables} (named list of data frames) and
\code{steps} (list of step definitions).
}
\description{
Shiny module server for PDF table extraction
}
