% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{show_steps}
\alias{show_steps}
\title{Show recorded steps}
\usage{
show_steps(sess)
}
\arguments{
\item{sess}{A \code{pdfmacro_session} object.}
}
\value{
\code{sess} invisibly.
}
\description{
Prints a numbered list of all steps in the session, with flagged steps
highlighted.
}
