% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{preview}
\alias{preview}
\title{Preview one extracted table}
\usage{
preview(sess, table, n = 6)
}
\arguments{
\item{sess}{A \code{pdfmacro_session} object.}

\item{table}{Character label of the table to preview.}

\item{n}{Number of rows to show (default 6).}
}
\value{
\code{sess} invisibly. # not recorded
}
\description{
Preview one extracted table
}
