% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{show_items}
\alias{show_items}
\title{Show all extracted items}
\usage{
show_items(sess)
}
\arguments{
\item{sess}{A \code{macrox_session} object.}
}
\value{
\code{sess} invisibly. # not recorded
}
\description{
Show all extracted items
}
